- Refactor de level selection view
