+++ app/LevelSelectionNavBarView.js
@@ -2,7 +2,7 @@ import React, {Component} from 'react';
 import {View, Text, StyleSheet} from 'react-native';
 import Icon from 'react-native-vector-icons/MaterialIcons';
 
-export default class ZoologicLevelSelectionNavBarView extends Component {
+export default class LevelSelectionNavBarView extends Component {
 
   constructor(props){
     super(props);
