+++ app/Navigation.js
@@ -8,13 +8,15 @@ import {Navigator} from 'react-native-deprecated-custom-components';
 import RushHourMainView from './rush-hour/RushHourMainView'
 import RushHourLevelView from './rush-hour/RushHourLevelView';
 import RushHourSuccessScreenView from './rush-hour/RushHourSuccessScreenView';
-import LevelSelectionView from "./rush-hour/LevelSelectionView";
 import ZoologicMainView from './zoologic/ZoologicMainView';
 import GameSelectionView from './rush-hour/GameSelectionView';
 import ZoologicLevelView from './zoologic/ZoologicLevelView';
 import ZoologicSuccessScreenView from './zoologic/ZoologicSuccessScreenView';
-import ZoologicLevelSelectionView from './zoologic/ZoologicLevelSelectionView';
+import LevelSelectionView from './LevelSelectionView';
 import zoologicStyle from './zoologic/ZoologicLevelSelectionStyle'
+import rushHourStyle from './rush-hour/RushHourLevelSelectionStyle'
+import ZoologicLevels from './zoologic/ZoologicLevels';
+import RushHourLevels from './rush-hour/RushHourLevels';
 
 export default class Navigation extends React.Component {
   constructor(props) {
@@ -24,14 +26,13 @@ export default class Navigation extends React.Component {
   render() {
     return (
       <Navigator
-        initialRoute={{id: 'zoologic-level-selection', levelNumber: 1}}
+        initialRoute={{id: 'rush-hour-level-selection', levelNumber: 1}}
         renderScene={this.navigatorRenderScene}
       />
     );
   }
 
   navigatorRenderScene(route, navigator) {
-    console.log(JSON.stringify(route));
     switch (route.id) {
       case 'game-selection':
         return (<GameSelectionView navigator={navigator} />);
@@ -43,7 +44,7 @@ export default class Navigation extends React.Component {
       case 'rush-hour-success-screen':
         return (<RushHourSuccessScreenView navigator={navigator} actualLevelNumber={route.levelNumber}/>);
       case 'rush-hour-level-selection':
-        return (<LevelSelectionView navigator={navigator}/>);
+        return (<LevelSelectionView levels={RushHourLevels.LEVELS} navigator={navigator} levelSelectionRouteId="rush-hour-lvl" backRouteId='rush-hour-main' style={rushHourStyle}/>);
 
       case 'zoologic-main':
         return (<ZoologicMainView navigator={navigator}/>);
@@ -52,7 +53,7 @@ export default class Navigation extends React.Component {
       case 'zoologic-success-screen':
         return (<ZoologicSuccessScreenView navigator={navigator} actualLevelNumber={route.levelNumber}/>);
       case 'zoologic-level-selection':
-        return (<ZoologicLevelSelectionView navigator={navigator} levelSelectionRouteId="zoologic-lvl" backRouteId='zoologic-main' style={zoologicStyle}/>);
+        return (<LevelSelectionView levels={ZoologicLevels.LEVELS} navigator={navigator} levelSelectionRouteId="zoologic-lvl" backRouteId='zoologic-main' style={zoologicStyle}/>);
     }
   }
